+++ examples/groq/scrape_plain_text_groq.py
@@ -33,6 +33,11 @@ graph_config = {
         "api_key": groq_key,
         "temperature": 0
     },
+     "embeddings": {
+        "model": "ollama/nomic-embed-text",
+        "temperature": 0,
+        # "base_url": "http://localhost:11434",  # set ollama URL arbitrarily
+    },
     "verbose": True,
     "headless": False
 }