+++ scrapegraphai/integrations/__init__.py
@@ -3,3 +3,4 @@ Init file for integrations module
 """
 
 from .burr_bridge import BurrBridge
+from .indexify_node import IndexifyNode