+++ scrapegraphai/nodes/generate_answer_node.py
@@ -11,11 +11,12 @@ from langchain_core.output_parsers import JsonOutputParser
 from langchain_core.runnables import RunnableParallel
 from tqdm import tqdm
 
-from ..utils.logging import get_logger
 
+from ..utils.logging import get_logger
+from ..models import Ollama
 # Imports from the library
 from .base_node import BaseNode
-from ..helpers import template_chunks, template_no_chunks, template_merge, template_chunks_with_schema, template_no_chunks_with_schema
+from ..helpers import template_chunks, template_no_chunks, template_merge
 
 
 class GenerateAnswerNode(BaseNode):
@@ -44,8 +45,12 @@ class GenerateAnswerNode(BaseNode):
         node_name: str = "GenerateAnswer",
     ):
         super().__init__(node_name, "node", input, output, 2, node_config)
-
+      
         self.llm_model = node_config["llm_model"]
+
+        if isinstance(node_config["llm_model"], Ollama):
+            self.llm_model.format="json"
+
         self.verbose = (
             True if node_config is None else node_config.get("verbose", False)
         )
@@ -76,42 +81,32 @@ class GenerateAnswerNode(BaseNode):
         user_prompt = input_data[0]
         doc = input_data[1]
 
-        output_parser = JsonOutputParser()
+        # Initialize the output parser
+        if self.node_config.get("schema", None) is not None:
+            output_parser = JsonOutputParser(pydantic_object=self.node_config["schema"])
+        else:
+            output_parser = JsonOutputParser()
+
         format_instructions = output_parser.get_format_instructions()
 
         chains_dict = {}
 
         # Use tqdm to add progress bar
         for i, chunk in enumerate(tqdm(doc, desc="Processing chunks", disable=not self.verbose)):
-            if self.node_config.get("schema", None) is None and len(doc) == 1:
+            if len(doc) == 1:
                 prompt = PromptTemplate(
                     template=template_no_chunks,
                     input_variables=["question"],
                     partial_variables={"context": chunk.page_content,
                                        "format_instructions": format_instructions})
-            elif self.node_config.get("schema", None) is not None and len(doc) == 1:
-                 prompt = PromptTemplate(
-                    template=template_no_chunks_with_schema,
-                    input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                       "format_instructions": format_instructions,
-                                       "schema": self.node_config.get("schema", None)
-                                       })
-            elif self.node_config.get("schema", None) is None and len(doc) > 1:
+                
+            else:
                 prompt = PromptTemplate(
                     template=template_chunks,
                     input_variables=["question"],
                     partial_variables={"context": chunk.page_content,
                                         "chunk_id": i + 1,
                                         "format_instructions": format_instructions})
-            elif self.node_config.get("schema", None) is not None and len(doc) > 1:
-                prompt = PromptTemplate(
-                    template=template_chunks_with_schema,
-                    input_variables=["question"],
-                    partial_variables={"context": chunk.page_content,
-                                        "chunk_id": i + 1,
-                                        "format_instructions": format_instructions,
-                                        "schema": self.node_config.get("schema", None)})
 
             # Dynamically name the chains based on their index
             chain_name = f"chunk{i+1}"